tests/wpt/web-platform-tests/bluetooth/service/getCharacteristics/characteristics-not-found.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getEmptyHealthThermometerService()
    .then(({service}) => assert_promise_rejects_with_message(
      service.getCharacteristics(),
      new DOMException('No Characteristics found in service.',
                       'NotFoundError')))
}, 'Request for absent characteristics. Reject with NotFoundError.');
</script>
